backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-bcrypt
pymysql
cryptography
Pillow
torch
ultralytics
